+++ src/pages/admin/order/single/[canvasOrderId].tsx
@@ -1,4 +1,4 @@
 import dynamic from 'next/dynamic';
-const AdminCanvasOrderDetail = dynamic(() => import('src/components/admin/canvasorder/AdminCanvasOrderDetail'));
+const AdminCanvasOrderDetail = dynamic(() => import('src/components/admin/order/AdminCanvasOrderDetail'));
 
 export default AdminCanvasOrderDetail;
