+++ src/pages/admin/order/single/index.tsx
@@ -1,4 +1,4 @@
 import dynamic from 'next/dynamic';
-const AdminCanvasOrderList = dynamic(() => import('src/components/admin/canvasorder/AdminCanvasOrderList'));
+const AdminCanvasOrderList = dynamic(() => import('src/components/admin/order/AdminCanvasOrderList'));
 
 export default AdminCanvasOrderList;
